mvp-tickets/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-filter
psycopg2-binary
Pillow
django-cors-headers
